Release/0.0.14

+++ src/main/java/org/circle8/dto/PuntoDto.java
@@ -4,8 +4,8 @@ import org.circle8.controller.response.PuntoResponse;
 import org.circle8.entity.Punto;
 
 public class PuntoDto {
-	public float latitud;
-	public float longitud;
+	public double latitud;
+	public double longitud;
 
 	public static PuntoDto from(Punto entity) {
 		if ( entity == null ) return null;
